now able to show the overview on shows
