+++ src/RightPanel.java
@@ -153,15 +153,17 @@ public class RightPanel extends JPanel implements View{
             if(movieListSelected()){
                 name = (String)movieModel.get(index);
                 JOptionPane.showMessageDialog(new JFrame(), "Movie Selected");
+                new MovieWindow(name);
+
                 //open url based on the movie name
             }
             else{
                 name = (String)showModel.get(index);
                 JOptionPane.showMessageDialog(new JFrame(), "Show Selected");
+                new ShowWindow(name);
                 //open url based on movie name
             }
             //create new window and frame based on the media selected
-            new MediaWindow(name);
         }
 
     }
